pyproject.toml
[tool.poetry]
name = "wikipy"
version = "0.1.0"
description = ""
authors = ["Owais Jamil <desertthunder.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.12.5"
rich = "^13.9.2"
httpx = "^0.27.2"
pydantic = "^2.9.2"
sqlmodel = "^0.0.22"
python-dotenv = "^1.0.1"
bs4 = "^0.0.2"
graphviz = "^0.20.3"


[tool.poetry.group.dev.dependencies]
ipython = "^8.28.0"
ruff = "^0.6.9"
coverage = "^7.6.1"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
